+++ packages/outstatic/src/app/api/index.tsx
@@ -4,24 +4,19 @@ import signout from '@/app/api/auth/signout'
 import user from '@/app/api/auth/user'
 import generate from '@/app/api/generate'
 import media from '@/app/api/media'
-import { Session } from 'next-session/lib/types'
 import { NextRequest } from 'next/server'
 
 export interface Request extends NextRequest {
-  session: Session
+  session: any
 }
 
-export type GetParams = {
-  params: {
-    ost: ['callback', 'login', 'signout', 'user', 'media']
-  }
-}
+export type GetParams = Promise<{
+  ost: ['callback', 'login', 'signout', 'user', 'media']
+}>
 
-export type PostParams = {
-  params: {
-    ost: ['generate']
-  }
-}
+export type PostParams = Promise<{
+  ost: ['generate']
+}>
 
 const getPaths = {
   callback,
@@ -36,13 +31,13 @@ const postPaths = {
 }
 
 export const OutstaticApi = {
-  GET: async (req: Request, { params }: GetParams) => {
-    const { ost } = params
+  GET: async (req: Request, segmentData: { params: GetParams }) => {
+    const { ost } = await segmentData.params
     const rsp = getPaths[ost[0]](req)
     return rsp
   },
-  POST: async (req: Request, { params }: PostParams) => {
-    const { ost } = params
+  POST: async (req: Request, segmentData: { params: PostParams }) => {
+    const { ost } = await segmentData.params
     const rsp = postPaths[ost[0]](req)
     return rsp
   }