+++ packages/outstatic/src/client/client.tsx
@@ -1,3 +1,7 @@
-export * from './pages'
-export * from './pages/settings/_components/media-settings'
-export * from '@/components/ui/outstatic/github-branch-search'
+export { AdminArea, Dashboard, OstClient } from './pages'
+export { MediaSettings } from './pages/settings/_components/media-settings'
+export { GitHubBranchSearch } from '@/components/ui/outstatic/github-branch-search'
+export {
+  useNavigationGuard,
+  NavigationGuardProvider
+} from 'next-navigation-guard'