+++ packages/outstatic/src/client/pages/collections/_components/new-collection-modal.tsx
@@ -12,11 +12,11 @@ import {
 import { Label } from '@/components/ui/shadcn/label'
 import { RadioGroup, RadioGroupItem } from '@/components/ui/shadcn/radio-group'
 import { createCommitApi } from '@/utils/createCommitApi'
-import { useCollections } from '@/utils/hooks'
+import { useCollections } from '@/utils/hooks/useCollections'
 import { useCreateCommit } from '@/utils/hooks/useCreateCommit'
 import { useGetDocuments } from '@/utils/hooks/useGetDocuments'
 import useOid from '@/utils/hooks/useOid'
-import useOutstatic from '@/utils/hooks/useOutstatic'
+import { useOutstatic } from '@/utils/hooks/useOutstatic'
 import { useRebuildMetadata } from '@/utils/hooks/useRebuildMetadata'
 import { zodResolver } from '@hookform/resolvers/zod'
 import Link from 'next/link'
@@ -465,12 +465,12 @@ export default function NewCollectionModal({
                           allowedChars: 'a-zA-Z0-9.'
                         })
                       : createFolder
-                        ? '/' +
-                          (path ? path + '/' : '') +
-                          slugify(form.getValues('name') || 'your-collection', {
-                            allowedChars: 'a-zA-Z0-9.'
-                          })
-                        : '/' + path
+                      ? '/' +
+                        (path ? path + '/' : '') +
+                        slugify(form.getValues('name') || 'your-collection', {
+                          allowedChars: 'a-zA-Z0-9.'
+                        })
+                      : '/' + path
                   }
                 />
                 <FormDescription>