+++ packages/outstatic/src/client/router.tsx
@@ -1,6 +1,6 @@
 import { AdminLoading } from '@/components/AdminLoading'
 import { useCollections } from '@/utils/hooks/useCollections'
-import useOutstatic from '@/utils/hooks/useOutstatic'
+import { useOutstatic } from '@/utils/hooks/useOutstatic'
 import { ReactElement } from 'react'
 import CustomFields from './pages/custom-fields'
 import Collections from './pages/collections'
@@ -8,6 +8,7 @@ import EditDocument from './pages/edit-document'
 import List from './pages/list'
 import Settings from './pages/settings'
 import MediaLibrary from './pages/media-library'
+import { EditorProvider } from '@/components/editor/editor-context'
 
 const defaultPages: { [key: string]: ReactElement | undefined } = {
   settings: <Settings />,
@@ -36,7 +37,9 @@ export const Router = ({ params }: { params: { ost: string[] } }) => {
       {!slug ? (
         <Collections />
       ) : slug2 && isContent ? (
-        <EditDocument collection={slug} />
+        <EditorProvider>
+          <EditDocument collection={slug} />
+        </EditorProvider>
       ) : !slug2 && isContent ? (
         <List slug={slug} title={title} />
       ) : slug === 'collections' &&