+++ packages/outstatic/src/utils/hooks/index.tsx
@@ -1,3 +1,3 @@
-export * from './useCollections'
-export * from './useInitialData'
-export * from './useOutstatic'
+export { useCollections } from './useCollections'
+export { useOutstatic, useLocalData } from './useOutstatic'
+export { InitialDataContext, useInitialData } from './useInitialData'